feat: modify the original function fetchPriceFromCoinGecko in order to receive a token symbol as parameter
