+++ packages/nextjs/utils/scaffold-stark/index.ts
@@ -1,5 +1,4 @@
 export * from "./networks";
 export * from "./notification";
 export * from "./fetchPriceFromCoingecko";
-export * from "./fetchStrkPriceFromCoingecko";
 export * from "./types";
